Move widgets from apputils to ui-components
