+++ packages/apputils/package.json
@@ -59,20 +59,17 @@
     "@lumino/disposable": "^1.4.3",
     "@lumino/domutils": "^1.2.3",
     "@lumino/messaging": "^1.4.3",
-    "@lumino/properties": "^1.2.3",
     "@lumino/signaling": "^1.4.3",
     "@lumino/virtualdom": "^1.8.0",
     "@lumino/widgets": "^1.19.0",
     "@types/react": "^17.0.0",
     "react": "^17.0.1",
-    "react-dom": "^17.0.1",
     "sanitize-html": "~2.3.3",
     "url": "^0.11.0"
   },
   "devDependencies": {
     "@jupyterlab/testutils": "^3.3.0-alpha.0",
     "@types/jest": "^26.0.10",
-    "@types/react-dom": "^17.0.0",
     "@types/sanitize-html": "^2.3.1",
     "jest": "^26.4.2",
     "rimraf": "~3.0.0",
