+++ packages/apputils/src/inputdialog.ts
@@ -1,9 +1,9 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
+import { Styling } from '@jupyterlab/ui-components';
 import { Widget } from '@lumino/widgets';
 import { Dialog, showDialog } from './dialog';
-import { Styling } from './styling';
 
 const INPUT_DIALOG_CLASS = 'jp-Input-Dialog';
