+++ packages/apputils/src/mainareawidget.ts
@@ -2,12 +2,11 @@
 // Distributed under the terms of the Modified BSD License.
 
 import { ITranslator, nullTranslator } from '@jupyterlab/translation';
+import { Spinner, Toolbar } from '@jupyterlab/ui-components';
 import { Message, MessageLoop } from '@lumino/messaging';
 import { BoxLayout, BoxPanel, Widget } from '@lumino/widgets';
 import { DOMUtils } from './domutils';
 import { Printing } from './printing';
-import { Spinner } from './spinner';
-import { Toolbar } from './toolbar';
 
 /**
  * A widget meant to be contained in the JupyterLab main area.
