+++ packages/ui-components/package.json
@@ -46,9 +46,13 @@
     "@blueprintjs/core": "^3.36.0",
     "@blueprintjs/select": "^3.15.0",
     "@jupyterlab/coreutils": "^5.3.0-alpha.0",
+    "@jupyterlab/translation": "^3.3.0-alpha.0",
     "@lumino/algorithm": "^1.3.3",
+    "@lumino/commands": "^1.12.0",
     "@lumino/coreutils": "^1.5.3",
     "@lumino/disposable": "^1.4.3",
+    "@lumino/messaging": "^1.4.3",
+    "@lumino/properties": "^1.2.3",
     "@lumino/signaling": "^1.4.3",
     "@lumino/virtualdom": "^1.8.0",
     "@lumino/widgets": "^1.19.0",
