+++ packages/ui-components/src/components/index.ts
@@ -1,7 +1,13 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
+export { Collapse as Collapser } from './collapse';
 export * from './htmlselect';
+export * from './iframe';
 export * from './interface';
 export * from './menu';
+export * from './spinner';
+export * from './styling';
 export * from './switch';
+export * from './toolbar';
+export * from './vdom';
